refactor: remove TerminalInfoProvider and use InkPictureProvider alias
